Added case for if file doesn't exist

+++ ExecuteMe/MoveCsv.ps1
@@ -1,6 +1,12 @@
 Function MoveCsv($FileName){
-    Remove-Item -Path "C:\xampp\mysql\data\webserver\$FileName.csv" -Force
-    Move-Item -Path "\\tiws07\dwg\Mfg Mtg\Customer Schedule\$FileName.csv" -Destination "C:\xampp\mysql\data\webserver\" -Force
+    $Path = "\\tiws07\dwg\Mfg Mtg\Customer Schedule\$FileName.csv"
+    if (Test-Path $Path -PathType Leaf){
+        Remove-Item -Path "C:\xampp\mysql\data\webserver\$FileName.csv" -Force
+        Move-Item -Path "\\tiws07\dwg\Mfg Mtg\Customer Schedule\$FileName.csv" -Destination "C:\xampp\mysql\data\webserver\" -Force
+        Write-Output "Moved $FileName to directory"
+    }else{
+        Write-Warning "$FileName does not exist"
+    }
 }
 $FileName = "Job Schedule Details"
 MoveCsv -File $FileName
